refactor: reworked app folder structure
