+++ src/app/[locale]/page.tsx
@@ -1,10 +1,8 @@
 import { JSX } from 'react';
 import Main from '@/components/main/Main';
-import { generateStaticParams } from '@/utils/utils';
+
 import { JikanApiResponse } from '@/types/types';
 import checkResponse from '@/utils/http/checkResponse';
-
-export { generateStaticParams };
 
 const MainPage = async (): Promise<JSX.Element> => {
   const page = 1;
